+++ src/components/checkbox/checkbox-group.vue
@@ -1,5 +1,9 @@
 <template>
-  <div>
+  <div
+    role="group"
+    :aria-label="ariaLabel"
+    :aria-disabled="disabled ? 'true' : null"
+  >
     <slot></slot>
   </div>
 </template>
@@ -12,8 +16,12 @@ export default {
       type: Array,
       required: true
     },
-    max: Number,
-    disabled: { type: Boolean }
+    max: {
+      type: Number,
+      validator: value => value >= 0
+    },
+    disabled: Boolean,
+    ariaLabel: String
   },
   watch: {
     value (newValue) {
@@ -23,17 +31,27 @@ export default {
   methods: {
     selectItem (item) {
       const { value } = this
-      if (this.max && this.value.length >= this.max) {
-        return
+      if (value.some(selectedItem => selectedItem === item)) {
+        return false
       }
+
+      if (this.max !== undefined && value.length >= this.max) return false
+
       this.$emit('input', [...value, item])
+      return true
     },
     deleteItem (item) {
       const { value: selectItems } = this
+
+      if (!selectItems.some(selectItem => selectItem === item)) {
+        return false
+      }
+
       this.$emit(
         'input',
         selectItems.filter(selectitem => selectitem !== item)
       )
+      return true
     }
   }
 }