+++ src/components/radio/radio.vue
@@ -1,21 +1,31 @@
 <template>
 <div>
-  <span
+  <label
     :class="[
       'radio-item',
       { 'is-checked': isChecked },
       { 'is-disabled': isDisabled },
       { 'is-group': isGroup }
     ]"
-    @click.stop="toggle"
+    @click.stop
   >
+    <input
+      v-bind="$attrs"
+      class="radio-native"
+      type="radio"
+      :name="nativeName"
+      :value="label"
+      :checked="isChecked"
+      :disabled="isDisabled"
+      @change="onChange"
+    >
     <span class="radio-inner">
       <bo-icon :name=iconName :size=iconSize :color=iconColor />
       <span class="slot-value">
         <slot></slot>
       </span>
     </span>
-  </span>
+  </label>
 </div>
 
 </template>
@@ -24,11 +34,13 @@
 import { findComponentUpward } from '../utils/assist'
 export default {
   name: 'bo-radio',
+  inheritAttrs: false,
   props: {
     value: {
       type: [String, Number]
     },
-    label: String,
+    label: [String, Number],
+    name: String,
     iconSize: {
       type: String,
       default: '24px'
@@ -63,14 +75,10 @@ export default {
       return (this.parent && this.parent.disabled) || this.disabled
     },
     isChecked () {
-      const { isGroup, model } = this
-      if (!isGroup) return model
-
-      const {
-        label,
-        parent: { value: selectItem }
-      } = this
-      return selectItem === label
+      return this.model === this.label
+    },
+    nativeName () {
+      return this.isGroup ? this.parent.nativeName : this.name
     },
     model: {
       get () {
@@ -86,8 +94,18 @@ export default {
     this.parent ? this.isGroup = true : this.isGroup = false
   },
   methods: {
-    toggle (event) {
-      !this.isDisabled && (this.model = this.label)
+    onChange (event) {
+      if (event.target.checked) this.select()
+    },
+    select () {
+      if (this.isDisabled || this.isChecked) return false
+
+      this.model = this.label
+      if (!this.isGroup) this.$emit('change', this.label)
+      return true
+    },
+    toggle () {
+      return this.select()
     }
   }
 }
@@ -96,13 +114,25 @@ export default {
 <style lang="stylus" scoped>
 @import '../../style/var';
 .radio-item
+  position relative
+  cursor pointer
   &.is-group
     display inline-block
     padding 2px 0
   &.is-checked
     color $theme-color
   &.is-disabled
+    cursor not-allowed
     color #CCC
+  .radio-native
+    position absolute
+    width 1px
+    height 1px
+    opacity 0
+    pointer-events none
+    &:focus + .radio-inner
+      outline 2px solid $theme-color
+      outline-offset 2px
   .radio-inner
     i,span
       vertical-align middle